// backend/src/features/conversation/domain/conversation.types.ts
import { type UserId } from '@/features/user/domain/user.types.js';
import { type Brand } from '@/types/brand.js';

export type ConversationId = Brand<string, 'ConversationId'>;

export interface Conversation {
  id: ConversationId;
  participants: UserId[]; // userIds
  isGroup: boolean;
  createdAt: Date;
  updatedAt: Date;
}